backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# LLM Providers
anthropic==0.18.1
openai==1.12.0
litellm==1.23.1

# Document Processing
python-docx==1.1.0
pypdf==4.0.1
PyMuPDF==1.23.26
pdf2image==1.17.0
pytesseract==0.3.10
Pillow==10.2.0

# Utilities
python-dotenv==1.0.1
aiofiles==23.2.1

# Development
pytest==8.0.0
pytest-asyncio==0.23.4
httpx==0.26.0
